move DTO logic inside the service layer

+++ src/main/java/org/green_building/excellent_training/services/UsersService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import org.green_building.excellent_training.dtos.UserDto;
 import org.green_building.excellent_training.entities.User;
 import org.green_building.excellent_training.repositories.RolesRepository;
 import org.green_building.excellent_training.repositories.UsersRepository;
@@ -20,26 +21,30 @@ public class UsersService {
         this.rolesRepository = rolesRepository;
     }
 
-    public List<User> getAll() {
-        return this.usersRepository.findAll();
+    public List<UserDto> getAll() {
+        List<User> users = this.usersRepository.findAll();
+        return UserDto.from(users);
     }
 
-    public User getById(Integer id) {
+    public UserDto getById(Integer id) {
         // findById(id) returns Optional<Role>
-        return this.usersRepository.findById(id).orElse(null);
+        User user = this.usersRepository.findById(id).orElse(null);
+        return UserDto.from(user);
     }
 
-    public User create(User user) {
-        if (user.getUsername() == null) return null;
-        if (user.getPassword() == null) return null;
-        if (user.getRole() == null) return null;
-        if (user.getRole().getId() == null) return null; // role id must be set
-        if (this.rolesRepository.findById(user.getRole().getId()).isEmpty()) return null; // role must exist in the db
-        return this.usersRepository.save(user);
+    public UserDto create(UserDto userDto) {
+        if (userDto.getUsername() == null) return null;
+        if (userDto.getPassword() == null) return null;
+        if (userDto.getRoleId()   == null) return null; // role id must be set
+        if (this.rolesRepository.findById(userDto.getRoleId()).isEmpty()) return null; // role must exist in the db
+        User user = User.from(userDto);
+        User createdUser = this.usersRepository.save(user);
+        return UserDto.from(createdUser);
     }
 
-    public User updateById(Integer id, User updates) {
-        User user = this.getById(id);
+    public UserDto updateById(Integer id, UserDto updatesDto) {
+        User updates = User.from(updatesDto);
+        User user = this.usersRepository.findById(id).orElse(null);
         if (user == null) return null;
         if (updates.getUsername() != null) user.setUsername(updates.getUsername());
         if (updates.getPassword() != null) user.setPassword(updates.getPassword());
@@ -49,18 +54,20 @@ public class UsersService {
             }
             user.setRole(updates.getRole());
         }
-        return this.usersRepository.save(user);
+        User updatedUser = this.usersRepository.save(user);
+        return UserDto.from(updatedUser);
     }
 
-    public List<User> deleteAll() {
+    public List<UserDto> deleteAll() {
         List<User> users = this.usersRepository.findAll();
         this.usersRepository.deleteAll();
-        return users;
+        return UserDto.from(users);
     }
 
-    public User deleteById(Integer id) {
-        User user = this.getById(id); // if it doesn't exist, then this will be null
+    public UserDto deleteById(Integer id) {
+        UserDto userDto = this.getById(id); // if it doesn't exist, then this will be null
+        if (userDto == null) return null;
         this.usersRepository.deleteById(id);
-        return user;
+        return userDto;
     }
 }
